util: Split up the commands into separate files in the m5 util.
This way each individual command can have a unit test written for it,
covering how it gathers its arguments and puts them passes them to the
underlying dispatch function.

Reviewed-on: https://gem5-review.googlesource.com/c/public/gem5/+/27564
Maintainer: Gabe Black <gabeblack@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>
